% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsplot_bar1.R
\name{jsplot.bar1}
\alias{jsplot.bar1}
\title{jsplot_bar1}
\usage{
jsplot.bar1(data, prob = FALSE, col = "dark blue", lab_x = "XNAME",
  lab_y = "YNAME", lab_main = "MAIN", xlab_size = 12,
  ylab_size = 12, main_size = 16, main_loc = "left",
  xtext_size = 10, table = TRUE, prob_table = FALSE, sort = TRUE)
}
\description{
jsplot_bar1
}
